ㄴㅁㅇ

+++ src/main/java/com/projectA1/repository/VisitCountingRepository.java
@@ -12,7 +12,7 @@ import com.projectA1.model.VisitCounting;
 public interface VisitCountingRepository extends JpaRepository<VisitCounting, Long> {
 
 	// userIdf를 기반으로 방문 횟수를 카운트하는 메서드
-	int countByUserId(Long userId);
+	Long countByUserId(Long userId);
 
 	VisitCounting findByUserAndCenter(User user, FitnessCenter center);
